requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.7
billiard==3.6.1.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
configparser==3.8.1
contextlib2==0.5.5
cryptography==2.7
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
Flask==1.1.1
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
idna==2.7
importlib-metadata==0.19
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
kombu==4.6.4
MarkupSafe==1.1.1
paramiko==2.6.0
pathlib2==2.3.4
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==3.13
redis==3.3.7
requests==2.20.1
requests-file==1.4.3
scandir==1.10.0
selenium==3.141.0
selenium-requests==1.3
six==1.12.0
SQLAlchemy==1.3.7
texttable==0.9.1
tldextract==2.2.1
typing==3.7.4
urllib3==1.24.3
vine==1.3.0
websocket-client==0.56.0
Werkzeug==0.15.5
zipp==0.5.2
